+++ test/lightning/jobs/job_test.exs
@@ -3,6 +3,8 @@ defmodule Lightning.Jobs.JobTest do
 
   alias Lightning.Jobs.Job
 
+  import Lightning.Factories
+
   defp random_job_name(length) do
     for _ <- 1..length,
         into: "",
@@ -13,6 +15,37 @@ defmodule Lightning.Jobs.JobTest do
   end
 
   describe "changeset/2" do
+    test "raises a constraint error when jobs in the same workflow have the same downcased and hyphenated name" do
+      workflow = insert(:workflow)
+
+      [first | rest] = [
+        "Validate form type",
+        "validate form type",
+        "validate-form-type",
+        "validate-FORM type"
+      ]
+
+      insert(:job, workflow: workflow, name: first)
+
+      Enum.each(rest, fn name ->
+        {:error, changeset} =
+          Job.changeset(
+            %Job{},
+            params_with_assocs(:job, workflow: workflow, name: name)
+          )
+          |> Repo.insert()
+
+        refute changeset.valid?
+
+        assert changeset.errors[:name] ==
+                 {"has already been taken",
+                  [
+                    constraint: :unique,
+                    constraint_name: "jobs_name_workflow_id_index"
+                  ]}
+      end)
+    end
+
     test "name can't be longer than 100 chars" do
       name = random_job_name(101)
       errors = Job.changeset(%Job{}, %{name: name}) |> errors_on()